It begins!  The modularization and refactorization
This is the start of a cleaving between the command line code, the
utilities, and the backend to the google calendar.

The idea is that eventually this should be modular and somewhat
more functional that what we started with.  In the long term, this will
make the code easier to test and develop.